+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.13 (unreleased)
 - Refactor GetAdventure/GetCompaign to a single method
+- Refactor to replace store/finished in Zone class with a DropReference
 - Bump saves library version
 
 ## v0.0.12 (28 May 2024)